sidtuneinfocache removed
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@154 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/applet/favorites/FavoritesModel.java
@@ -35,7 +35,7 @@ public class FavoritesModel extends DefaultTableModel {
 			Enum.class, Enum.class, Enum.class, Long.class, String.class,
 			Integer.class, Integer.class, Integer.class, Integer.class,
 			Integer.class, Integer.class, Integer.class, Date.class,
-			Integer.class, Integer.class, Integer.class, Integer.class };
+			Integer.class, Integer.class, Short.class, Short.class };
 
 	public static final String HVSC_PREFIX = "<HVSC>";
